能否更进一步qwq

+++ src/Interaction.cpp
@@ -1,6 +1,7 @@
 #include "Interaction.h"
 #include "Defs.h"
 #include "Logic.h"
+#include <cstdint>
 #include <iostream>
 
 using namespace std;
@@ -18,12 +19,16 @@ int initAndRestore(Board &board) {
     if (x0 == -1)
       myColor = grid_black;
     else {
-      Move move = {{x0, y0}, {x1, y1}, {x2, y2}};
+      Move move = {{static_cast<int8_t>(x0), static_cast<int8_t>(y0)},
+                   {static_cast<int8_t>(x1), static_cast<int8_t>(y1)},
+                   {static_cast<int8_t>(x2), static_cast<int8_t>(y2)}};
       Logic::applyMove(board, move, -myColor);
     }
     if (i <= turnID - 1) {
       cin >> x0 >> y0 >> x1 >> y1 >> x2 >> y2;
-      Move move = {{x0, y0}, {x1, y1}, {x2, y2}};
+      Move move = {{static_cast<int8_t>(x0), static_cast<int8_t>(y0)},
+                   {static_cast<int8_t>(x1), static_cast<int8_t>(y1)},
+                   {static_cast<int8_t>(x2), static_cast<int8_t>(y2)}};
       Logic::applyMove(board, move, myColor);
     }
   }
